# Conflicts:
#	.gitignore
#	src/integration-test/java/org/web3j/protocol/core/CoreIT.java
#	src/integration-test/java/org/web3j/protocol/scenarios/DeployContractIT.java
#	src/integration-test/java/org/web3j/protocol/scenarios/FunctionWrappersIT.java
#	src/integration-test/java/org/web3j/protocol/scenarios/HumanStandardTokenGeneratedIT.java
#	src/integration-test/java/org/web3j/protocol/scenarios/HumanStandardTokenIT.java
#	src/main/java/org/web3j/abi/TypeDecoder.java
#	src/main/java/org/web3j/codegen/AbiTypesGenerator.java
#	src/main/java/org/web3j/codegen/AbiTypesMapperGenerator.java
#	src/main/java/org/web3j/console/Runner.java
#	src/main/java/org/web3j/console/WalletRunner.java
#	src/main/java/org/web3j/crypto/Wallet.java
#	src/main/java/org/web3j/protocol/core/methods/response/Transaction.java
#	src/main/java/org/web3j/protocol/http/HttpService.java
#	src/main/java/org/web3j/protocol/infura/InfuraHttpService.java
#	src/main/java/org/web3j/protocol/rx/JsonRpc2_0Rx.java
#	src/main/java/org/web3j/tx/Contract.java
#	src/test/java/org/web3j/TempFileProvider.java
#	src/test/java/org/web3j/codegen/SolidityFunctionWrapperGeneratorTest.java
#	src/test/java/org/web3j/protocol/core/RequestTest.java
#	src/test/java/org/web3j/protocol/core/filters/FilterTester.java

+++ src/integration-test/java/org/web3j/protocol/scenarios/DeployContractIT.java
@@ -9,7 +9,6 @@ import org.web3j.abi.FunctionEncoder;
 import org.web3j.abi.FunctionReturnDecoder;
 import org.web3j.abi.datatypes.Function;
 import org.web3j.abi.datatypes.Type;
-import org.web3j.abi.datatypes.Uint;
 import org.web3j.protocol.core.DefaultBlockParameterName;
 import org.web3j.protocol.core.methods.request.Transaction;
 import org.web3j.protocol.core.methods.response.TransactionReceipt;
@@ -17,9 +16,9 @@ import org.web3j.protocol.core.methods.response.TransactionReceipt;
 import static junit.framework.TestCase.assertFalse;
 import static org.hamcrest.core.Is.is;
 import static org.hamcrest.core.IsEqual.equalTo;
+import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.assertNotNull;
 
 /**
  * Integration test demonstrating the full contract deployment workflow.
@@ -43,6 +42,7 @@ public class DeployContractIT extends Scenario {
                 transactionReceipt.getGasUsed().equals(GAS_LIMIT));
 
         String contractAddress = transactionReceipt.getContractAddress();
+
         assertNotNull(contractAddress);
 
         Function function = createFibonacciFunction();
@@ -53,7 +53,7 @@ public class DeployContractIT extends Scenario {
         List<Type> uint = FunctionReturnDecoder.decode(
                 responseValue, function.getOutputParameters());
         assertThat(uint.size(), is(1));
-        assertThat((BigInteger) uint.get(0).getValue(), equalTo(BigInteger.valueOf(13)));
+        assertThat(uint.get(0).getValue(), equalTo(BigInteger.valueOf(13)));
     }
 
     private String sendTransaction() throws Exception {
