+++ src/main/java/org/web3j/protocol/core/methods/response/EthBlock.java
@@ -14,8 +14,8 @@ import com.fasterxml.jackson.databind.ObjectReader;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 
 import org.web3j.protocol.ObjectMapperFactory;
-import org.web3j.utils.Numeric;
 import org.web3j.protocol.core.Response;
+import org.web3j.utils.Numeric;
 
 /**
  * Block object returned by:
@@ -402,7 +402,7 @@ public class EthBlock extends Response<EthBlock.Block> {
         public TransactionObject(String hash, String nonce, String blockHash, String blockNumber,
                                  String transactionIndex, String from, String to, String value,
                                  String gasPrice, String gas, String input, String creates,
-                                 String publicKey, String raw, String r, String s, byte v) {
+                                 String publicKey, String raw, String r, String s, int v) {
             super(hash, nonce, blockHash, blockNumber, transactionIndex, from, to, value,
                     gasPrice, gas, input, creates, publicKey, raw, r, s, v);
         }
